fixed integer bitwise shift

+++ src/radSens1v2.cpp
@@ -11,10 +11,9 @@ ClimateGuard_RadSens1v2::~ClimateGuard_RadSens1v2()
 {
 }
 
-
 /*Initialization function and sensor connection. Returns false if the sensor is not connected to the I2C bus.*/
 bool ClimateGuard_RadSens1v2::radSens_init()
-{   
+{
 #if defined(ARDUINO)
     Wire.begin();
     Wire.beginTransmission(_sensor_address); //safety check, make sure the sensor is connected
@@ -22,10 +21,11 @@ bool ClimateGuard_RadSens1v2::radSens_init()
     if (Wire.endTransmission(true) != 0)
         return false;
 #elif defined(__arm__)
-	_fd = wiringPiI2CSetup(_sensor_address);
-    if (_fd == -1) {
+    _fd = wiringPiI2CSetup(_sensor_address);
+    if (_fd == -1)
+    {
         return false;
-	}
+    }
 #endif
     updateData();
     return true;
@@ -35,12 +35,14 @@ bool ClimateGuard_RadSens1v2::updateData()
 {
 #if defined(ARDUINO)
     Wire.requestFrom(_sensor_address, 19);
-    for (int i = 0; i < 19; i++) {
+    for (int i = 0; i < 19; i++)
+    {
         _data[i] = Wire.read();
     }
 #elif defined(__arm__)
     uint8_t reg_adr = 0x00;
-    for (int i = 0; i < 19; i++) {
+    for (int i = 0; i < 19; i++)
+    {
         _data[i] = wiringPiI2CReadReg8(_fd, reg_adr);
         reg_adr += 1;
     }
@@ -57,8 +59,8 @@ bool ClimateGuard_RadSens1v2::getData()
 {
     if (updateData())
     {
-        intensy_static = ((_data[6] << 16) | (_data[7] << 8) | _data[8]) / 10.0;
-        intensy_dyanmic = ((_data[3] << 16) | (_data[4] << 8) | _data[5]) / 10.0;
+        intensy_static = (((uint32_t)_data[6] << 16) | ((uint16_t)_data[7] << 8) | _data[8]) / 10.0;
+        intensy_dyanmic = (((uint32_t)_data[3] << 16) | ((uint16_t)_data[4] << 8) | _data[5]) / 10.0;
         pulses = _pulse_cnt;
         return true;
     }
@@ -84,7 +86,7 @@ float ClimateGuard_RadSens1v2::getRadIntensyDyanmic()
 {
     if (updateData())
     {
-        float temp = ((_data[3] << 16) | (_data[4] << 8) | _data[5]) / 10.0;
+        float temp = (((uint32_t)_data[3] << 16) | ((uint16_t)_data[4] << 8) | _data[5]) / 10.0;
         return temp;
     }
     else
@@ -98,7 +100,7 @@ float ClimateGuard_RadSens1v2::getRadIntensyStatic()
 {
     if (updateData())
     {
-        return ((_data[6] << 16) | (_data[7] << 8) | _data[8]) / 10.0;
+        return (((uint32_t)_data[6] << 16) | ((uint16_t)_data[7] << 8) | _data[8]) / 10.0;
     }
     else
     {
@@ -136,10 +138,12 @@ bool ClimateGuard_RadSens1v2::getHVGeneratorState()
 {
     if (updateData())
     {
-        if(_data[17] == 1) {
+        if (_data[17] == 1)
+        {
             return true;
         }
-        else {
+        else
+        {
             return false;
         }
     }
@@ -165,27 +169,39 @@ bool ClimateGuard_RadSens1v2::setHVGeneratorState(bool state)
 {
 #if defined(ARDUINO)
     Wire.beginTransmission(0x66);
-    #if (ARDUINO >= 100)
-        Wire.write(RS_HV_GENERATOR_RG);    
-        if (state) {
-            Wire.write(1);
-        } else {
-            Wire.write(0);  
-        }
-    #else
-        Wire.send(RS_HV_GENERATOR_RG);   
-        if (state) {
-            Wire.send(1);
-        } else {
-            Wire.send(0);  
-        }
-    #endif
-    if (Wire.endTransmission(true) == 0) return true;  //"true" sends stop message after transmission & releases I2C bus
+#if (ARDUINO >= 100)
+    Wire.write(RS_HV_GENERATOR_RG);
+    if (state)
+    {
+        Wire.write(1);
+    }
+    else
+    {
+        Wire.write(0);
+    }
+#else
+    Wire.send(RS_HV_GENERATOR_RG);
+    if (state)
+    {
+        Wire.send(1);
+    }
+    else
+    {
+        Wire.send(0);
+    }
+#endif
+    if (Wire.endTransmission(true) == 0)
+        return true; //"true" sends stop message after transmission & releases I2C bus
 #elif defined(__arm__)
-    if (state) {
-        if (wiringPiI2CWriteReg8(_fd, RS_HV_GENERATOR_RG, 1) > 0) return true;
-    } else {
-        if (wiringPiI2CWriteReg8(_fd, RS_HV_GENERATOR_RG, 0) > 0) return true;
+    if (state)
+    {
+        if (wiringPiI2CWriteReg8(_fd, RS_HV_GENERATOR_RG, 1) > 0)
+            return true;
+    }
+    else
+    {
+        if (wiringPiI2CWriteReg8(_fd, RS_HV_GENERATOR_RG, 0) > 0)
+            return true;
     }
 #endif
     return false;
@@ -202,16 +218,18 @@ bool ClimateGuard_RadSens1v2::setSensitivity(uint8_t sens)
 {
 #if defined(ARDUINO)
     Wire.beginTransmission(0x66);
-    #if (ARDUINO >= 100)
-        Wire.write(RS_SENSITIVITY_RG);
-        Wire.write(sens);
-    #else
-        Wire.send(RS_SENSITIVITY_RG);
-        Wire.send(sens);
-    #endif
-    if (Wire.endTransmission(true) == 0) return true; 
+#if (ARDUINO >= 100)
+    Wire.write(RS_SENSITIVITY_RG);
+    Wire.write(sens);
+#else
+    Wire.send(RS_SENSITIVITY_RG);
+    Wire.send(sens);
+#endif
+    if (Wire.endTransmission(true) == 0)
+        return true;
 #elif defined(__arm__)
-    if (wiringPiI2CWriteReg8(_fd, RS_SENSITIVITY_RG, sens) > 0) return true;
+    if (wiringPiI2CWriteReg8(_fd, RS_SENSITIVITY_RG, sens) > 0)
+        return true;
 #endif
     return false;
 }
